Base UI extension UIColor [addFile] = [UIColor+.swift]
