+++ testCode/testCode/testCode/ViewController.swift
@@ -666,20 +666,6 @@ class MyComponent: GKComponent {
  * クラスに対し、インスタンスをアーカイブする際にNSKeyedArchiverやNSKeyedUnarchiverで利用される名前を指定します
  */
 
-extension UIColor {
-    convenience init(hex: String, alpha: CGFloat) {
-        let v = hex.map { String($0) } + Array(repeating: "0", count: max(6 - hex.count, 0))
-        let r = CGFloat(Int(v[0] + v[1], radix: 16) ?? 0) / 255.0
-        let g = CGFloat(Int(v[2] + v[3], radix: 16) ?? 0) / 255.0
-        let b = CGFloat(Int(v[4] + v[5], radix: 16) ?? 0) / 255.0
-        self.init(red: r, green: g, blue: b, alpha: min(max(alpha, 0), 1))
-    }
-    
-    convenience init(hex: String) {
-        self.init(hex: hex, alpha: 1.0)
-    }
-}
-
 class Monster: NSObject {
     var name: String
     @objc dynamic var hitPoint: Int
